Bottom App bar design done
