+++ app/build.gradle
@@ -165,4 +165,7 @@ dependencies {
     implementation "io.coil-kt:coil:2.1.0"
     implementation "io.coil-kt:coil-compose:2.1.0"
 
+    //ICONS
+    implementation "androidx.compose.material:material-icons-extended:$compose_version"
+
 }
