--- app/src/main/java/com/alphaomardiallo/parisforkids/ui/activity/MainActivity.kt
@@ -1,54 +1,0 @@
-package com.alphaomardiallo.parisforkids.ui.activity
-
-import android.content.ContentValues.TAG
-import android.os.Bundle
-import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.viewModels
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.alphaomardiallo.parisforkids.ui.theme.ParisForKidsTheme
-import com.alphaomardiallo.parisforkids.viewmodel.MainViewModel
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-
-    private val viewModel: MainViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            val list = viewModel.eventsAndActivities
-            Log.i(TAG, "onCreate: ${list.toString()}")
-            ParisForKidsTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colors.background
-                ) {
-                    Greeting("Android")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun Greeting(name: String) {
-    Text(text = "Hello $name!")
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    ParisForKidsTheme {
-        Greeting("Android")
-    }
-}
