+++ app/src/main/java/com/alphaomardiallo/parisforkids/ui/activity/main/MainCompose.kt
@@ -1,0 +1,2 @@
+package com.alphaomardiallo.parisforkids.ui.activity.main
+
